+++ pos_longpolling/static/src/js/pos.js
@@ -1,0 +1,42 @@
+odoo.define('pos_longpolling.pos', function(require){
+    "use strict";
+
+    var models = require('point_of_sale.models');
+    var PosLongpolling = require('pos_longpolling.connection');
+
+    var PosModelSuper = models.PosModel;
+    models.PosModel = models.PosModel.extend({
+        initialize: function() {
+            var self = this;
+            PosModelSuper.prototype.initialize.apply(this, arguments);
+            this.buses = {};
+            var options = {
+                name: 'bus_service',
+                lonpolling_activated: false
+            };
+            this.bus = new PosLongpolling(options);
+            this.ready.then(function () {
+                self.bus.init_bus(self);
+                if (self.config.autostart_longpolling) {
+                    self.bus.start();
+                }
+            });
+        },
+        add_bus: function(key, sync_server) {
+            var options = {
+                name: key,
+                route: sync_server
+            };
+            this.buses[key] = new PosLongpolling(options);
+            this.buses[key].init_bus(this);
+        },
+        get_bus: function(key){
+            if (key){
+                return this.buses[key];
+            }
+            return this.bus;
+        },
+    });
+
+    return models;
+});
