+++ pos_longpolling/static/src/js/sync_bus_service.js
@@ -1,0 +1,17 @@
+odoo.define('pos_longpolling.SyncBusService', function (require) {
+    "use strict";
+
+    var BusService = require('bus.BusService');
+
+    var SyncBusService = BusService.extend({
+        _isActive: true,
+        addPollRoute: function(url) {
+            if (url) {
+                this.POLL_ROUTE = url + this.POLL_ROUTE;
+            }
+        }
+    });
+
+    return SyncBusService;
+
+});
